Move top level layer canvas components to separate directory.
