+++ src/ui/layout/HorizontalLayout.js
@@ -9,8 +9,8 @@ import PubSub from 'pubsub-js';
 import {dataLoaded, mapAdded, mapRemoved, mapReorder, reset, featureUpdate} from '../../topics';
 import {LayoutBase} from './LayoutBase';
 import {Bounds} from '../../model/Bounds';
-import {BioMap as BioMapComponent} from '../../canvas/layout/BioMap';
-import {CorrespondenceMap as CorrMapComponent} from '../../canvas/layout/CorrespondenceMap';
+import {BioMap as BioMapComponent} from '../../canvas/canvas/BioMap';
+import {CorrespondenceMap as CorrMapComponent} from '../../canvas/canvas/CorrespondenceMap';
 import {QtlTrack} from '../../canvas/layout/QtlTrack';
 import {Popover} from '../menus/Popover';
 import {FeatureMenu} from '../menus/FeatureMenu';
